pyproject.toml
[build-system]
requires = ["maturin==1.9.2"]
build-backend = "maturin"

[tool.maturin]
manifest-path = "./al-rust/Cargo.toml"

[project]
name = "approximate-lexical"
version = "0.1.0"
description = "Approximate lexical search using Rust and Python, testing against BM25s."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bm25s[full]>=0.2.13",
    "hnswlib>=0.8.0",
    "matplotlib>=3.10.5",
    "pystemmer>=3.0.0",
    "rich>=14.1.0",
    "torch>=2.7.1",
    "typer>=0.16.0",
]